% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{calculate_pairwise_distances_matrix}
\alias{calculate_pairwise_distances_matrix}
\title{Calculate Pairwise Distance Matrix}
\usage{
calculate_pairwise_distances_matrix(
  data,
  x_col = "x",
  y_col = "y",
  distance_fun = distGeo
)
}
\arguments{
\item{data}{A data frame containing site coordinates and IDs.}

\item{x_col}{Name of the longitude column in \code{data} (default: "x").}

\item{y_col}{Name of the latitude  column in \code{data} (default: "y").}

\item{distance_fun}{The distance function to use (default: \code{distGeo}).}
}
\value{
A data frame containing pairwise distances between sites (in km).
}
\description{
Calculate Pairwise Distance Matrix
}
\examples{
# default usage when your coords are in columns "x" and "y"
my_sites_df <- data.frame(
  grid_id = 1:4,
  x = c(18.4, 18.6, 19.0, 19.2),
  y = c(-33.9, -34.0, -33.7, -33.5)
)
dist_df_default <- calculate_pairwise_distances_matrix(data = my_sites_df)
head(dist_df_default)

# specify custom column names for longitude and latitude
sites_custom <- data.frame(
  grid_id = 1:4,
  centroid_lon = c(18.4, 18.6, 19.0, 19.2),
  centroid_lat = c(-33.9, -34.0, -33.7, -33.5)
)
dist_df <- calculate_pairwise_distances_matrix(
  data  = sites_custom,
  x_col = "centroid_lon",
  y_col = "centroid_lat"
)
head(dist_df)
}
\keyword{internal}
