% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{calculate_pairwise_gower_dist_matrix}
\alias{calculate_pairwise_gower_dist_matrix}
\title{Calculate Pairwise Gower Dissimilarity Matrix}
\usage{
calculate_pairwise_gower_dist_matrix(df, sp_cols)
}
\arguments{
\item{df}{A data frame containing site information.}

\item{sp_cols}{A vector of column names for species data.}
}
\value{
A data frame containing pairwise Gower dissimilarities between sites.
}
\description{
Calculate Pairwise Gower Dissimilarity Matrix
}
