% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{cor_pears}
\alias{cor_pears}
\title{Calculate Pearson's Correlation}
\usage{
cor_pears(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A numeric vector representing species abundances at the first site.}

\item{vec_to}{A numeric vector representing species abundances at the second site.}
}
\value{
Pearson's correlation coefficient.
}
\description{
Pearson's correlation for linear abundance relationships; aggregated across sites.
}
