% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissmapr-package.R
\docType{package}
\name{dissmapr-package}
\alias{dissmapr}
\alias{dissmapr-package}
\title{dissmapr: Workflow for Compositional Dissimilarity and Biodiversity Turnover Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to compute and map biodiversity turnover at the macroscale. 'dissmapr' provides a reproducible, end-to-end workflow for computing and mapping compositional dissimilarity and biodiversity turnover across large spatial scales, including occurrence-data retrieval, environmental-data extraction, flexible grid generation, order-wise dissimilarity computation (zeta diversity and multi-site generalised dissimilarity modelling) and bioregional mapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/dissmapr}
  \item \url{https://b-cubed-eu.github.io/dissmapr/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/dissmapr/issues}
}

}
\author{
\strong{Maintainer}: Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})

Authors:
\itemize{
  \item Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})
}

Other contributors:
\itemize{
  \item Stellenbosch University [copyright holder]
  \item European Union (Horizon Europe grant agreement ID 101059592 (B-Cubed)) [funder]
}

}
\keyword{internal}
