% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissmapr-package.R
\docType{_PACKAGE}
\name{dissmapr}
\alias{dissmapr}
\title{dissmapr: Workflow for Compositional Dissimilarity & Biodiversity Turnover Analysis}
\description{
\code{dissmapr} is an R package for analysing compositional dissimilarity and biodiversity turnover across spatial gradients.
}
