% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{distance}
\alias{distance}
\title{Calculate Distance Between Sites}
\usage{
distance(df, site_col, vec_from, vec_to = NULL)
}
\arguments{
\item{df}{A data frame containing site coordinates.}

\item{site_col}{The column name representing site IDs.}

\item{vec_from}{The site ID or vector for the starting site.}

\item{vec_to}{The site ID or vector for the destination site(s).}
}
\value{
Distance(s) in meters between the specified sites.
}
\description{
Calculate Distance Between Sites
}
