% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_df.R
\name{format_df}
\alias{format_df}
\title{Format Data Frame}
\usage{
format_df(
  data,
  format = NULL,
  x_col = NULL,
  y_col = NULL,
  site_id_col = NULL,
  species_col = NULL,
  value_col = NULL,
  sp_col_range = NULL,
  extra_cols = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} of biodiversity records.}

\item{format}{Character; one of \code{"long"} or \code{"wide"}. If \code{NULL}, inferred
from presence of \code{species_col} & \code{value_col}.}

\item{x_col}{Name of longitude column. If \code{NULL}, will search common names.}

\item{y_col}{Name of latitude column. If \code{NULL}, will search common names.}

\item{site_id_col}{Name of site identifier column. If \code{NULL}, will be
generated from \verb{(x, y)}.}

\item{species_col}{Name of species column (required for \code{"long"}).}

\item{value_col}{Name of value column (e.g. presence/abundance; for \code{"long"}).}

\item{sp_col_range}{Integer vector indexing species columns (for \code{"wide"}).}

\item{extra_cols}{Character vector of other columns to keep (e.g. metadata).}
}
\value{
A \code{list} with elements:
\itemize{
\item \code{site_obs}: data frame in \strong{long} format (only for \code{format="long"}).
\item \code{site_sp}: data frame in \strong{wide} format (species as columns).
}
}
\description{
This function formats biodiversity data into long or wide format for
further analysis.
}
