% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{gower_dissimilarity}
\alias{gower_dissimilarity}
\title{Calculate Gower's Dissimilarity}
\usage{
gower_dissimilarity(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A numeric or categorical vector for the first site.}

\item{vec_to}{A numeric or categorical vector for the second site.}
}
\value{
The Gower dissimilarity value.
}
\description{
Calculate Gower's Dissimilarity
}
