% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{mutual_info}
\alias{mutual_info}
\title{Calculate Mutual Information}
\usage{
mutual_info(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A numeric or categorical vector for the first variable.}

\item{vec_to}{A numeric or categorical vector for the second variable.}
}
\value{
The mutual information value between the two variables.
}
\description{
Calculate Mutual Information
}
