% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{phi_coef}
\alias{phi_coef}
\title{Calculate Phi Coefficient}
\usage{
phi_coef(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A binary presence-absence vector for the first site.}

\item{vec_to}{A binary presence-absence vector for the second site.}
}
\value{
The Phi Coefficient value.
}
\description{
Phi coefficient for presence-absence; -1 to +1 pairwise, summary of means/variances across combinations.
}
