% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{richness}
\alias{richness}
\title{Calculate Species Richness}
\usage{
richness(vec_from, vec_to = NULL)
}
\arguments{
\item{vec_from}{A numeric vector representing species counts at the first site.}

\item{vec_to}{(Optional) A numeric vector for pairwise or higher-order comparisons.}
}
\value{
The species richness value.
}
\description{
Calculate Species Richness
}
