% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{turnover}
\alias{turnover}
\title{Calculate Species Turnover or Beta Diversity}
\usage{
turnover(vec_from, vec_to = NULL)
}
\arguments{
\item{vec_from}{A numeric vector representing species counts at the first site.}

\item{vec_to}{(Optional) A numeric vector for pairwise or higher-order comparisons.}
}
\value{
The species turnover value.
}
\description{
Computes beta diversity as proportion unshared; 0-1 for pairwise, extended to multi-site averages or totals.
}
