% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{calculate_pairwise_distances_matrix}
\alias{calculate_pairwise_distances_matrix}
\title{Calculate Pairwise Distance Matrix}
\usage{
calculate_pairwise_distances_matrix(data, distance_fun = distGeo)
}
\arguments{
\item{data}{A data frame containing site coordinates and IDs.}

\item{distance_fun}{The distance function to use (default: \code{distGeo}).}
}
\value{
A data frame containing pairwise distances between sites.
}
\description{
Calculate Pairwise Distance Matrix
}
