% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{cor_spear}
\alias{cor_spear}
\title{Calculate Spearman's Rank Correlation}
\usage{
cor_spear(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A numeric vector representing species abundances at the first site.}

\item{vec_to}{A numeric vector representing species abundances at the second site.}
}
\value{
Spearman's rank correlation coefficient.
}
\description{
Calculate Spearman's Rank Correlation
}
