% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\name{diss_bcurt}
\alias{diss_bcurt}
\title{Calculate Bray-Curtis Dissimilarity}
\usage{
diss_bcurt(vec_from, vec_to)
}
\arguments{
\item{vec_from}{A numeric vector representing species abundances at the first site.}

\item{vec_to}{A numeric vector for species abundances at the second site.}
}
\value{
The Bray-Curtis dissimilarity value.
}
\description{
Calculate Bray-Curtis Dissimilarity
}
