% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enviro_data.R
\name{get_enviro_data}
\alias{get_enviro_data}
\title{Get Environmental Data}
\usage{
get_enviro_data(
  data,
  buffer_km = 10,
  source = "geodata",
  var = "bio",
  res = 2.5,
  path = "data/",
  year = NULL,
  depth = NULL,
  stat = "mean",
  model = NULL,
  ssp = NULL,
  time = NULL,
  temporal = FALSE,
  time_col = "year",
  layer_split_char = NULL,
  sp_cols = NULL,
  ext_cols = NULL
)
}
\arguments{
\item{data}{A data frame containing spatial points with x and y coordinates.}

\item{buffer_km}{Numeric. Buffer distance around the input points in kilometers (default: 10).}

\item{source}{Character. Data source: "geodata" for downloading or "local" for local rasters (default: "geodata").}

\item{var}{Character. Variable to retrieve (e.g., "bio" for bioclimatic variables, "elev" for elevation).}

\item{res}{Numeric. Resolution of the environmental data in arc-minutes (default: 2.5).}

\item{path}{Character. Directory path to store downloaded data or to load local rasters (default: "data/").}

\item{year}{Numeric. Year for temporal data (e.g., for population or footprint datasets; required for specific datasets).}

\item{depth}{Character. Depth range for soil data (e.g., "0-5cm").}

\item{stat}{Character. Statistic to summarize soil data (e.g., "mean").}

\item{model}{Character. Climate model for CMIP6 projections (e.g., "ACCESS-CM2").}

\item{ssp}{Numeric. Shared Socioeconomic Pathway for CMIP6 (e.g., 245 for SSP2-4.5).}

\item{time}{Character. Time period for CMIP6 projections (e.g., "2041-2060").}
}
\value{
A list containing:
\itemize{
\item \code{env_rast}: Raster of the environmental data for the area of interest.
\item \code{sites_sf}: Spatial points as an sf object.
\item \code{env_df}: Data frame combining input data and extracted environmental variables.
}
}
\description{
This function retrieves environmental data for spatial points, either by downloading data from
specified sources (e.g., WorldClim, SoilGrids) or using local raster files. Data is extracted for
a buffer region around the input points, and missing values are interpolated using surrounding values.
}
\examples{
# Example usage:
data = data.frame(site_id = 1:5, x = c(10, 12, 14, 16, 18), y = c(20, 22, 24, 26, 28))
env_data = get_enviro_data(data, buffer_km = 5, var = "bio", res = 2.5, path = "data/")
plot(env_data$env_rast[[1]])
points(env_data$sites_sf)
}
