% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_indices.R
\docType{data}
\name{required_packages}
\alias{required_packages}
\title{Helper Functions for \code{compute_orderwise}}
\format{
An object of class \code{character} of length 5.
}
\usage{
required_packages
}
\description{
These functions are designed to be used as the \code{func} parameter in the \code{compute_orderwise} function.
They calculate various metrics, including distances, dissimilarities, correlations, and mutual information.
}
\keyword{datasets}
